ValidationViewModelaBase implementiert für Errors in ViewModels

+++ CYR/Invoice/InvoiceRepositorys/InvoiceRepository.cs
@@ -170,6 +170,7 @@ namespace CYR.Invoice.InvoiceRepositorys
             decimal? nettAmount = 0;
             decimal? grossAmount = 0;
             bool succes = false;
+            if (invoice.Items.Any(i => i.HasErrors)) return succes;
             await _databaseConnection.ExecuteTransactionAsync(async (transaction) =>
             {
                 string deletePositionsQuery = "delete from Rechnungspositionen where Rechnungsnummer = @Rechnungsnummer";
